+++ src/gdal_exp.cpp
@@ -245,7 +245,7 @@ bool create(std::string format, std::string dst_filename,
 //' An error is raised if the operation fails.
 //' @seealso
 //' [`GDALRaster-class`][GDALRaster], [create()], [rasterFromRaster()],
-//' [getCreationOptions()]
+//' [getCreationOptions()], [translate()]
 //' @examples
 //' lcp_file <- system.file("extdata/storm_lake.lcp", package="gdalraster")
 //' tif_file <- paste0(tempdir(), "/", "storml_lndscp.tif")
@@ -991,6 +991,80 @@ bool sieveFilter(std::string src_filename, int src_band,
 }
 
 
+//' Convert raster data between different formats
+//'
+//' `translate()` is a wrapper of the \command{gdal_translate} command-line
+//' utility (see \url{https://gdal.org/programs/gdal_translate.html}).
+//' The function can be used to convert raster data between different
+//' formats, potentially performing some operations like subsetting,
+//' resampling, and rescaling pixels in the process. Refer to the GDAL
+//' documentation at the URL above for a list of command-line arguments that
+//' can be passed in `cl_arg`.
+//'
+//' @param src_filename Character string. Filename of the source raster.
+//' @param dst_filename Character string. Filename of the output raster.
+//' @param cl_arg Optional character vector of command-line arguments for 
+//' \code{gdal_translate}.
+//' @returns Logical indicating success (invisible \code{TRUE}).
+//' An error is raised if the operation fails.
+//' 
+//' @seealso
+//' [`GDALRaster-class`][GDALRaster], [rasterFromRaster()], [warp()]
+//'
+//' @examples
+//' # convert the elevation raster to Erdas Imagine format and resample to 90m
+//' elev_file <- system.file("extdata/storml_elev.tif", package="gdalraster")
+//'
+//' # command-line arguments for gdal_translate
+//' args <- c("-tr", "90", "90", "-r", "average")
+//' args <- c(args, "-of", "HFA", "-co", "COMPRESSED=YES")
+//'
+//' img_file <- paste0(tempdir(), "/", "storml_elev_90m.img")
+//' translate(elev_file, img_file, args)
+//' 
+//' ds <- new(GDALRaster, img_file, read_only=TRUE)
+//' ds$getDriverLongName()
+//' ds$bbox()
+//' ds$res()
+//' ds$getStatistics(band=1, approx_ok=FALSE, force=TRUE)
+//' ds$close()
+// [[Rcpp::export(invisible = true)]]
+bool translate(std::string src_filename, std::string dst_filename,
+		Rcpp::Nullable<Rcpp::CharacterVector> cl_arg = R_NilValue) {
+
+	GDALDatasetH src_ds = GDALOpenShared(src_filename.c_str(), GA_ReadOnly);
+	if (src_ds == NULL)
+		Rcpp::stop("Open source raster failed.");
+
+	std::vector<char *> argv = {NULL};
+	if (cl_arg.isNotNull()) {
+		// cast Nullable to the underlying type
+		Rcpp::CharacterVector cl_arg_in(cl_arg);
+		argv.resize(cl_arg_in.size() + 1);
+		for (R_xlen_t i = 0; i < cl_arg_in.size(); ++i) {
+			argv[i] = (char *) (cl_arg_in[i]);
+		}
+		argv[cl_arg_in.size()] = NULL;
+	}
+	
+	GDALTranslateOptions* psOptions = GDALTranslateOptionsNew(argv.data(), NULL);
+	if (psOptions == NULL)
+		Rcpp::stop("Translate failed (could not create options struct).");
+	GDALTranslateOptionsSetProgress(psOptions, GDALTermProgressR, NULL);
+	
+	GDALDatasetH hDstDS = GDALTranslate(dst_filename.c_str(), src_ds,
+							psOptions, NULL);
+							
+	GDALTranslateOptionsFree(psOptions);
+	GDALClose(src_ds);
+	if (hDstDS == NULL)
+		Rcpp::stop("Translate raster failed.");
+	
+	GDALClose(hDstDS);
+	return true;
+}
+
+
 //' Raster reprojection and mosaicing
 //'
 //' `warp()` is a wrapper of the \command{gdalwarp} command-line utility for
@@ -1140,7 +1214,7 @@ bool sieveFilter(std::string src_filename, int src_band,
 //' inputs to the SRS of `src_files[1]` when they are different).
 //'
 //' @seealso
-//' [`GDALRaster-class`][GDALRaster], [srs_to_wkt()]
+//' [`GDALRaster-class`][GDALRaster], [srs_to_wkt()], [translate()]
 //'
 //' @examples
 //' # reproject the elevation raster to NAD83 / CONUS Albers (EPSG:5070)